moved query into hook. closes to working so far with infinite scroll + lazy load.
