+++ src/components/Card.js
@@ -1,10 +1,13 @@
 import React, { useState, useRef, useEffect } from "react";
+import UseBirdSearch from "../UseBirdSearch";
 import GetMap from "./GetMap";
 import ImageModal from "./ImageModal";
 
 const Card = (props) => {
   const [isOpen, setIsOpen] = useState(false);
   const [isSeen, setIsSeen] = useState(false);
+
+  // UseBirdSearch(query);
 
   const cardRef = useRef();
   const options = {
@@ -14,7 +17,7 @@ const Card = (props) => {
 
   useEffect(() => {
     const observer = new IntersectionObserver((entries) => {
-      entries.forEach((entry) => {
+      entries.forEach((entry, index) => {
         if (isSeen && !entry.isIntersecting) {
           entry.target.classList.remove("card__show");
         } else {
@@ -23,7 +26,7 @@ const Card = (props) => {
           setIsSeen(true);
         }
       });
-    }, options);
+    });
 
     observer.observe(cardRef.current);
   });
